LightUp Logic
There seems to be an issue, however.
Something seems to be running asynchronously as all of the selected images will light up at the same time and then later darken again. Not sure why this is occurring. Check console logs for what I mean.

Goal: Have each light up to have in sequence (as the game would be played) instead of all at once.